+++ apps/prairielearn/src/ee/lib/lti13.ts
@@ -25,8 +25,8 @@ import {
   Lti13InstanceSchema,
   UserSchema,
 } from '../../lib/db-types.js';
-import { features } from '../../lib/features/index.js';
 import { type ServerJob } from '../../lib/server-jobs.js';
+import { selectUsersWithCourseInstanceAccess } from '../../models/course-instances.js';
 import { selectLti13Instance } from '../models/lti13Instance.js';
 
 import { getInstitutionAuthenticationProviders } from './institution.js';
@@ -333,12 +333,6 @@ export class Lti13Claim {
 export async function validateLti13CourseInstance(
   resLocals: Record<string, any>,
 ): Promise<boolean> {
-  const feature_enabled = await features.enabledFromLocals('lti13', resLocals);
-
-  if (!feature_enabled) {
-    return false;
-  }
-
   const hasLti13CourseInstance = await queryRow(
     sql.select_ci_validation,
     {
@@ -673,10 +667,10 @@ export const Lti13ScoreSchema = z.object({
 });
 export type Lti13Score = z.infer<typeof Lti13ScoreSchema>;
 
-const UsersWithLti13SubSchema = UserSchema.extend({
+const UserWithLti13SubSchema = UserSchema.extend({
   lti13_sub: z.string().nullable(),
 });
-type UsersWithLti13Sub = z.infer<typeof UsersWithLti13SubSchema>;
+type UserWithLti13Sub = z.infer<typeof UserWithLti13SubSchema>;
 
 // https://www.imsglobal.org/spec/lti-nrps/v2p0/#sharing-of-personal-data
 const ContextMembershipSchema = z.object({
@@ -696,17 +690,20 @@ const ContextMembershipContainerSchema = z.object({
 });
 
 class Lti13ContextMembership {
-  #memberships: Record<string, ContextMembership[]> = {};
+  #membershipsByEmail: Record<string, ContextMembership[]> = {};
+  #membershipsBySub: Record<string, ContextMembership> = {};
 
   private constructor(memberships: ContextMembership[]) {
     // Turn array into an object for efficient lookups. We need to retain duplicates
     // so that we can detect and handle the case where two users have the same email.
     for (const member of memberships) {
+      this.#membershipsBySub[member.user_id] = member;
+
       if (member.email === undefined) {
         continue;
       }
-      this.#memberships[member.email] ??= [];
-      this.#memberships[member.email].push(member);
+      this.#membershipsByEmail[member.email] ??= [];
+      this.#membershipsByEmail[member.email].push(member);
     }
   }
 
@@ -751,27 +748,25 @@ class Lti13ContextMembership {
   }
 
   /**
-   * @param user The user to look up.
-   * @returns The LTI 1.3 sub (user_id) for the user, or null if not found.
+   * @param user The user to look up with optional lti13_sub
+   * @returns The LTI 1.3 record for the user, or null if not found.
    */
-  async lookup(user: UsersWithLti13Sub): Promise<string | null> {
+  lookup(user: UserWithLti13Sub): ContextMembership | null {
     if (user.lti13_sub !== null) {
-      return user.lti13_sub;
+      return this.#membershipsBySub[user.lti13_sub] ?? null;
     }
-
     for (const match of ['uid', 'email']) {
-      const memberResults = this.#memberships[user[match]];
+      const memberResults = this.#membershipsByEmail[user[match]];
 
       if (!memberResults) continue;
 
       // member.email cannot be duplicated in memberships
       if (memberResults.length > 1) return null;
 
-      // The `user_id` that we get from the membership API is what we call `lti13_sub`.
-      return memberResults[0].user_id;
+      return memberResults[0];
     }
 
-    // The user wan't found.
+    // The user wasn't found.
     return null;
   }
 }
@@ -799,9 +794,7 @@ export async function updateLti13Scores(
     throw new HttpStatusError(403, 'Invalid assessment.id');
   }
 
-  job.info(`Sending grade data for ${assessment.tid} ${assessment.title}`);
-
-  const token = await getAccessToken(instance.lti13_instance.id);
+  job.info(`Working on assessment ${assessment.title} (${assessment.tid})`);
 
   const assessment_instances = await queryRows(
     sql.select_assessment_instances_for_scores,
@@ -811,22 +804,56 @@ export async function updateLti13Scores(
     AssessmentInstanceSchema.extend({
       score_perc: z.number(), // not .nullable() from SQL query
       date: DateFromISOString, // not .nullable() from SQL query
-      users: UsersWithLti13SubSchema.array(),
+      users: UserWithLti13SubSchema.array(),
     }),
   );
 
+  const courseStaff = await selectUsersWithCourseInstanceAccess({
+    course_instance_id: assessment.course_instance_id,
+    minimal_role: 'Student Data Viewer',
+  });
+  const courseStaffUids = new Set(courseStaff.map((staff) => staff.uid));
+
   const memberships = await Lti13ContextMembership.loadForInstance(instance);
+
   const timestamp = new Date();
+  const counts = {
+    success: 0,
+    error: 0,
+    not_sent: 0,
+  };
 
   for (const assessment_instance of assessment_instances) {
     for (const user of assessment_instance.users) {
-      job.info(`ai=${assessment_instance.id}, ${assessment_instance.score_perc}% for ${user.name}`);
+      // Get/Refresh the token in the main loop in case it expires during the run.
+      const token = await getAccessToken(instance.lti13_instance.id);
 
-      const userId = await memberships.lookup(user);
-      if (userId === null) {
-        job.warn(`* Could not find LTI user information for ${user.name} ${user.uid}, skipping...`);
+      const ltiUser = memberships.lookup(user);
+      const isCourseStaff = courseStaffUids.has(user.uid);
+
+      // User not found in LTI, reporting only
+      if (ltiUser === null) {
+        job.info(
+          `Not sending grade ${assessment_instance.score_perc.toFixed(2)}% for ${user.uid}.` +
+            ` Could not find ${isCourseStaff ? 'course staff' : 'student'} ${user.uid}` +
+            ` in ${instance.lti13_instance.name} course ${instance.lti13_course_instance.context_label}`,
+        );
+        counts.not_sent++;
         continue;
       }
+
+      // User is not a student in LTI, reporting only
+      if (!ltiUser.roles.includes(STUDENT_ROLE)) {
+        job.info(
+          `Not sending grade ${assessment_instance.score_perc.toFixed(2)}% for ${user.uid}.` +
+            ` ${isCourseStaff ? 'Course staff' : 'Student'} ${user.uid} is not a student` +
+            ` in ${instance.lti13_instance.name} course ${instance.lti13_course_instance.context_label}`,
+        );
+        counts.not_sent++;
+        continue;
+      }
+
+      job.info(`Sending grade ${assessment_instance.score_perc.toFixed(2)}% for ${user.uid}.`);
 
       const submittedAt = await selectAssessmentInstanceLastSubmissionDate(assessment_instance.id);
 
@@ -841,7 +868,7 @@ export async function updateLti13Scores(
         scoreMaximum: 100,
         activityProgress: assessment_instance.open ? 'Submitted' : 'Completed',
         gradingProgress: 'FullyGraded',
-        userId,
+        userId: ltiUser.user_id,
         submission: {
           startedAt: assessment_instance.date,
           submittedAt: submittedAt ?? undefined,
@@ -857,10 +884,26 @@ export async function updateLti13Scores(
         body: JSON.stringify(score),
       });
 
-      job.info(`\t${res.statusText}`);
-      if (!res.ok) {
-        job.warn(`\t${await res.text()}`);
+      if (res.ok) {
+        counts.success++;
+      } else {
+        counts.error++;
+        const results = await res.json();
+
+        // Default to showing the whole error
+        job.warn(`\t${res.statusText}`);
+        job.warn(`\t${JSON.stringify(results)}`);
       }
     }
   }
+  job.info('Done.\n\nSummary:');
+  job.info(`${counts.success} score${counts.success === 1 ? '' : 's'} successfully sent.`);
+  job.info(`${counts.error} score${counts.error === 1 ? '' : 's'} not sent due to errors.`);
+  if (counts.error > 0 && counts.success === 0) {
+    job.warn('\tNo scores successfully sent and errors are present.');
+    job.warn(
+      `\tIs the ${instance.lti13_instance.name} ${instance.lti13_course_instance.context_label} course published?`,
+    );
+  }
+  job.info(`${counts.not_sent} score${counts.not_sent === 1 ? '' : 's'} skipped (not sent).`);
 }