+++ apps/prairielearn/src/lib/db-types.ts
@@ -20,11 +20,14 @@ export const AssessmentsFormatForQuestionSchema = z.array(
   }),
 );
 
-const JsonCommentSchema = z.union([z.string(), z.array(z.any()), z.record(z.any())]);
+export const JsonCommentSchema = z.union([z.string(), z.array(z.any()), z.record(z.any())]);
 
 // *******************************************************************************
 // Enum schemas. These should be alphabetized by their corresponding enum name.
 // *******************************************************************************
+
+export const EnumGradingMethodSchema = z.enum(['Internal', 'External', 'Manual']);
+export type EnumGradingMethod = z.infer<typeof EnumGradingMethodSchema>;
 
 export const EnumJobStatusSchema = z.enum(['Running', 'Success', 'Error']);
 export type EnumJobStatus = z.infer<typeof EnumJobStatusSchema>;
@@ -101,6 +104,7 @@ export const AssessmentSchema = z.object({
   duration_stat_threshold_seconds: z.number().array(),
   duration_stat_thresholds: IntervalSchema.array(),
   group_work: z.boolean().nullable(),
+  honor_code: z.string().nullable(),
   id: IdSchema,
   json_comment: JsonCommentSchema.nullable(),
   json_grade_rate_minutes: z.number().nullable(),
@@ -346,6 +350,7 @@ export const CourseInstanceSchema = z.object({
   id: IdSchema,
   json_comment: JsonCommentSchema.nullable(),
   long_name: z.string().nullable(),
+  share_source_publicly: z.boolean(),
   short_name: z.string().nullable(),
   sync_errors: z.string().nullable(),
   sync_job_sequence_id: IdSchema.nullable(),
@@ -543,6 +548,7 @@ export const GroupConfigSchema = z.object({
   maximum: z.number().nullable(),
   minimum: z.number().nullable(),
   name: z.string().nullable(),
+  student_authz_choose_name: z.boolean(),
   student_authz_create: z.boolean().nullable(),
   student_authz_leave: z.boolean().nullable(),
   student_authz_join: z.boolean().nullable(),
@@ -823,7 +829,7 @@ export const QuestionSchema = z.object({
   external_grading_files: z.any().nullable(),
   external_grading_image: z.string().nullable(),
   external_grading_timeout: z.number().nullable(),
-  grading_method: z.enum(['Internal', 'External', 'Manual']),
+  grading_method: EnumGradingMethodSchema,
   id: IdSchema,
   json_comment: JsonCommentSchema.nullable(),
   json_external_grading_comment: z