+++ apps/prairielearn/src/tests/chunks.test.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 
-import { assert } from 'chai';
 import fs from 'fs-extra';
 import * as tmp from 'tmp-promise';
+import { afterEach, assert, beforeEach, describe, it } from 'vitest';
 import { z } from 'zod';
 
 import * as sqldb from '@prairielearn/postgres';
@@ -139,7 +139,7 @@ describe('chunks', () => {
       );
     });
 
-    it('should identify complex assessment in simple course instance', () => {
+    it('should identify complex assessment in complex course instance', () => {
       const chunks = chunksLib.identifyChunksFromChangedFiles(
         [
           'courseInstances/complex/course/instance/assessments/complex/assessment/clientFilesAssessment/file.txt',
@@ -260,9 +260,7 @@ describe('chunks', () => {
     });
   });
 
-  describe('ensureChunksForCourse', function () {
-    this.timeout(60000);
-
+  describe('ensureChunksForCourse', { timeout: 60_000 }, function () {
     let tempTestCourseDir: tmp.DirectoryResult;
     let tempChunksDir: tmp.DirectoryResult;
     const originalChunksConsumerDirectory = config.chunksConsumerDirectory;
@@ -272,7 +270,7 @@ describe('chunks', () => {
     let questionId;
     let nestedQuestionId;
 
-    beforeEach('set up testing server', async () => {
+    beforeEach(async () => {
       // We need to modify the test course - create a copy that we can
       // safely manipulate.
       tempTestCourseDir = await tmp.dir({ unsafeCleanup: true });
@@ -293,7 +291,7 @@ describe('chunks', () => {
       config.chunksConsumerDirectory = tempChunksDir.path;
       config.chunksConsumer = true;
 
-      await helperServer.before(tempTestCourseDir.path).call(this);
+      await helperServer.before(tempTestCourseDir.path)();
 
       // Find the ID of this course
       const results = await sqldb.queryOneRowAsync(sql.select_course_by_path, {
@@ -326,14 +324,14 @@ describe('chunks', () => {
       nestedQuestionId = nestedQuestionResults.rows[0].id;
     });
 
-    afterEach('shut down testing server', async () => {
+    afterEach(async () => {
       try {
         await tempTestCourseDir.cleanup();
         await tempChunksDir.cleanup();
       } catch (err) {
         console.error(err);
       }
-      await helperServer.after.call(this);
+      await helperServer.after();
 
       config.chunksConsumer = false;
       config.chunksConsumerDirectory = originalChunksConsumerDirectory;