+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prairielearn/postgres-tools
 
+## 2.0.18
+
+### Patch Changes
+
+- Updated dependencies [d97b97a]
+  - @prairielearn/postgres@2.1.14
+
+## 2.0.17
+
+### Patch Changes
+
+- be4444e: Upgrade all JavaScript dependencies
+- Updated dependencies [be4444e]
+  - @prairielearn/postgres@2.1.13
+
 ## 2.0.16
 
 ### Patch Changes